+++ lib/src/dtos/schema_field.g.dart
@@ -12,7 +12,6 @@ SchemaField _$SchemaFieldFromJson(Map<String, dynamic> json) => SchemaField(
       type: json['type'] as String? ?? "",
       system: json['system'] as bool? ?? false,
       required: json['required'] as bool? ?? false,
-      unique: json['unique'] as bool? ?? false,
       options: json['options'] as Map<String, dynamic>? ?? const {},
     );
 
@@ -23,6 +22,5 @@ Map<String, dynamic> _$SchemaFieldToJson(SchemaField instance) =>
       'type': instance.type,
       'system': instance.system,
       'required': instance.required,
-      'unique': instance.unique,
       'options': instance.options,
     };