+++ test/dtos/collection_model_test.dart
@@ -17,7 +17,6 @@ void main() {
             "type": "schema_type",
             "system": true,
             "required": false,
-            "unique": false,
             "options": {"a": 123},
           },
         ],